requirements.txt
semver==3.0.1
gitpython==3.1.32
